routescanner: Scrape direct connections
Notebook to scrape all the direct connections data from Routescanner, and save it as a dataframe. It currently finds about 170 connections between sea terminals in South America and The Netherlands/Belgium.

For each route, the company, origin, destination, service code, duration and if known frequency are saved.

A Pandas DataFrame is generated and saved in CSV and Pickle files.

LIMITATIONS: Currently it scrapes at most 20 results per route, because other results are on 2nd, 3rd etc pages.