Routescanner: Improved sleeping behaviour
Improved the sleeping behaviour of the routescanner web scraper. It now sleeps longer and restarts fewer. It also stops if iterations or sleep times get ridiculously long, and then returns the collected data.

Also fixed a depreciation issue with the Selenium webdriver.Chrome(), now inputting a service instead of a executable_path